Fix bugs regarding dentries crossing a block boundary.
1. "if statement" in add_dentry() function.
If a new dentry fits the edge of inode block exactly (c.f. *offset + rec_len == 4096 * n),
add_dentry() function regarded it as a crossing one even though we can place it in the current inode block.
It was obviously a bug because this behavior did not conform to the behavior of dentry_size() function.

So, I added an additional expression ((*offset + rec_len) % info.block_size) to the if statement.
Now the code detects the crossing condition of a new dentry very correctly in any circumstance.

2. padding extra dentry at the end of make_directory() function.
Padding an extra dentry of EXT4_FT_UNKNOWN type is necessary in order to complete the current inode block.
However, we have to consider the possibility that this padding dentry may exceed the current inode size.
In such a case, the revised code just expand the length of previous dentry instead of adding a new dentry.

Change-Id: I466f026c2c6f65e92e38067804559024caeb62a1