Extend make_ext4fs to support setting SELinux security contexts in ext4 images.
Extend make_ext4fs with support for looking up the right security context from
the file_contexts configuration and setting it in the generated image.  This is
similar to the existing support for looking up the UID/GID/mode via
android_filesystem_config.h and setting it, but via configuration rather than
defined in a header.

Change-Id: Ief9c44eeaaca4a44100a384b063f40b185469be3