Extend make_ext4fs() interface to allow callers to pass selabel_handle.
Extend make_ext4fs() to allow callers to pass an selabel_handle for
labeling files in the ext4 image.  Previously, this was only done
via the _internal() function.  This extends the library interface
so that it can be used by the recovery and updater code for labeling
files from OTA and update packages.

Change-Id: I4f6755fe7c65b69587276d17ef5b971ebec0161f