Update patch set 1
Patch Set 1:

There is compression in gmail (it's sent up to gmail servers via a private protocol, then actually sent via SMTP from the datacenter), but not necessarily for other types of accounts (IMAP/POP/Exchange).

It will mean (for gmail users at least) that you have to download the attachment and decompress it; you won't have the option to view it in the browser.  I don't have a strong feeling about this either way.

Patch-set: 1