Update patch set 3
Patch Set 3:

It's not related to different pagesizes. The issue was with a particular framebuffer driver where the kernel would report a value for smem_start that was a small value (something like 0x140 from memory). I think the first part of the page was reserved.

Benoit was the originator of the code and may be able to provide some additional information.

Patch-set: 3
Reviewer: Gerrit User 1003284 <1003284@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0