make_ext4fs: fix dentry padding when dentry size is 4088 or 4092
When the total dentry size is N * 4096 - 4 or N * 4096 - 8, there is
not enough room in the block to insert an extra padding dentry.
Instead, the last dentry should be increased in size to fill the block.
Modify make_directory to increase the size of the last dentry in this
case.

Also fix dentry_size to return the size of all the dentries
without the padding, the len + 8 padding is incorrect, and the
DIV_ROUND_UP will round up to the correct number of blocks.

Change-Id: If11f90e4ee172c135a0eae2a50b8be213e2dd0dc