Update patch set 2
Patch Set 2: Code-Review+1

I would prefer this code always be included, but that it is called if either an environment variable is set appropriately or if a command-line argument is used to enable it.

I don't mind the current patch, though, so marking it as approved.

Patch-set: 2
Reviewer: Gerrit User 1010466 <1010466@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1