Update patch set 1
Patch Set 1:

Note to reviewers: we've been using this for a while in a successor to Diskinstaller that we've developed. I think this code is generic enough that it may be useful in other scenarios, so I submit here as a library.

Patch-set: 1
Reviewer: Gerrit User 1011333 <1011333@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0