Update patch set 6
Patch Set 6:

would it make more sense to put these in system/netd? i feel like tests are more likely to be run the closer they are to the code they test.

(i've started moving the bionic tests out of here. not least because the "real" bionic tests are part of CTS which makes them immensely more valuable.)

Patch-set: 6