Label lost+found directories with SELinux security contexts.
When an asec image is generated, the lost+found directory has no
security.selinux attribute set and thus shows up as unlabeled.
We already handle this correctly when building images from existing
directories with files, so we just need to do it here as well when
generating an otherwise empty initial image.

Change-Id: I6d449dcc67a5cbc2101718e34ea0e184936d7f0e
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>