Update patch set 2
Patch Set 2:

sorry for not replying sooner... the plan with these tests is to go through them all and make sure that the tests in bionic/tests/ test everything that's tested here. if something's already tested in bionic (which means it's in CTS too!), we can just remove the corresponding test here. if the test here is useless, we can just delete it. if the test here is useful and not already covered by bionic's tests, we should improve bionic's tests and then remove the test here.

i've been working on this when i get time, but haven't had change to remove any lately.

in the meantime, if you just want to silence build warnings, i'm fine with just removing the Android.mk files.

Patch-set: 2
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0