Move make_ext4fs_internal() declaration fo make_ext4fs.h
The make_ext4fs() function calls make_ext4fs_internal() with a
|wipe| argument set to 1, which always fails horribly on non Linux
systems. This is an issue when trying to use the library within
the emulator.

Changing this default might have consequences on other parts of the
build, so work-around the issue by exposing the raw make_ext4fs_internal()
in the public library header instead.

BUG=none

Change-Id: I62b62b7c3e99b465c3b876154231e7c2fe541b23