Ignore wipe operation on non-block devices.
The implementation of make_ext4fs() calls make_ext4fs_internal() by
forcing the |wipe| parameter to true, which is problematic when the
library is used within the emulator (the wipe operation will always
fail on non-Linux platforms).

The wipe is only useful when writing directly to a block device,
so only call wipe_block_device() when it is certain that the file
descriptor points to a block device.

Also convert the fatal error into a warning, when trying to actually
wipe a block device on non-Linux systems.

BUG=NONE

Change-Id: I62b62b7c3e99b465c3b876154231e7c2fe541b23