Create make_ext4fs_extra() to disable verbose and wipe.
The implementation of make_ext4fs() calls make_ext4fs_internal() by
forcing the |wipe| parameter to true, which is problematic when the
library is used within the emulator (the wipe operation will always
fail on non-Linux platforms).

To work-around this, introduce make_ext4fs_extra() which allows
the caller to disable this parameter explicitly. Also allow the
caller to enable |verbose| for easier debugging.

BUG=NONE

Change-Id: I62b62b7c3e99b465c3b876154231e7c2fe541b23