bionic libc test: clean up pthread mutex related test
since most of the tests have been re-implemented in file
    bionic/tests/pthread_test.cpp
in the gtest format as tests:
   pthread.pthread_mutex_lock_NORMAL
   pthread.pthread_mutex_lock_ERRORCHECK
   pthread.pthread_mutex_lock_RECURSIVE

so clean up the files and settings here with this change.

One thing here is that The test for PTHREAD_PROCESS_SHARED
and the complex test case do not exist in the new implementation.

Change-Id: Iee414692d97cc85bc491214af131f7e988b6bc8b
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>