Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

I've looked at the failure, and it is due to a trivial conflict with the internal version of this file that is very easy to fix. I don't think there is a way to fix this here, so I'll handle it.

Patch-set: 1
Reviewer: Gerrit User 1000411 <1000411@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Verified=+1