Update patch set 1
Patch Set 1:

I've added the comment. The bug is present in all glibc versions so I cannot activate the code based on __GLIBC__/__GLIBC_MINOR__ values. It is simply guarded by a ifdef __GLIBC__. I think the emulator has a copy of this file so the change should be applied there too.

Patch-set: 1
Label: Verified=0