Add a program to display a slideshow at early boot
Add a small program to display a slideshow of images (specified
on the command line) on the framebuffer device. Images are switched
after a timeout, or user input. The timeout can be optionally set
using the -t command line parameter and defaults to 5s. Pressing the
power button exits the slideslow immediately.

Once the last image is reached, the program will wait for the power
button to be pressed before exiting. If the button is not pressed in
five minutes, the program exits to avoid draining the battery during
an accidental reboot.

Change-Id: I58c79a7f3ac747eec0d73a10f018d3d8ade9df7d
(Moved from I743c68b5f38038abcaaf5ae08233fa72fd823c8e)