Update patch set 1
Patch Set 1:

(2 comments)

As JP pointed out, this may not be an appropriate warning method for all devices, so the call will be in device specific init.rc.

The hash mentioned in the design document applies to the warning shown by the bootloader, where the hash of the user provided key used to sign the boot image should be shown.

Patch-set: 1
Label: Verified=0