Update patch set 3
Patch Set 3: Verified+1

for the record, sizeof and variable length arrays do work. the real bug was that argc is the number of non-null elements in the array. i prefer the code without the memset anyway, and it fails safer.

Patch-set: 3
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1