Update patch set 2
Patch Set 2:

(3 comments)

This looks great -- really good progress. It is marked as "WIP" so I'm not sure whether you want low-level comments yet.

One feature to which I would like to call attention is the collection of build IDs -- this is a critical requirement for my use case (android-wide profiling), since we can't do the offline symbolization without it. I don't think it's a lot of work to implement, but I wanted to make sure you had it on your radar.

I pulled your changes, built simplerperf, and tried running it on my N5. The resulting perf.data looks fine at first glance except for the fact that there doesn't seem to be a dummy MMAP for the kernel, which one typically sees in perf.data files. 

Thanks, Than

Patch-set: 2
Label: Verified=0