Update patch set 1
Patch Set 1:

I don't think this is right. The ext4 code in the kernel which parses this xattr block doesn't have the same pointer to the end of the struct. We can't succeed just because we reached the end of some boundary...

Of course, I may not understand how the kernel parses this block completely. It may be easier to talk in person to make sure we both understand what's going on.

Patch-set: 1
Label: Code-Review=0