Update patch set 1
Patch Set 1: Code-Review-1

Then the _problem_ is that we managed to get within sizeof(uint32_t) of the value. We are missing the 'four null bytes' when deciding to move on to the next block, in the case of the failure I found, the 'capabilities' were butted against the entries.

The documentation agrees with you in fs/ext4/xattr.c:

 * The attributes in inodes and on blocks have a different header; the entries
 * are stored in the same format:
 *
 *   +------------------+
 *   | header           |
 *   | entry 1          | |
 *   | entry 2          | | growing downwards
 *   | entry 3          | v
 *   | four null bytes  |
 *   | . . .            |
 *   | value 1          | ^
 *   | value 3          | | growing upwards
 *   | value 2          | |
 *   +------------------+

Patch-set: 1
Label: Code-Review=-1