ext4_utils: write superblocks at last
Because ext4_utils uses lazy itable init by default, if the file
system creation is interrupted on an already-have-data partition,
it will be very likely to have a mixed file system, where the old
data may still be valid, which is not expected.

To reduce the possibility, zero superblocks are written to the disk
along with other file-system metadata in a first pass, and backup
superblocks are written to the disk in a 2nd pass, and the primary
superblock at the very last. Anything written to the disk before the
valid superblocks are written will result in an invalid file-system,
and will trigger a new file-system creation or factory reset.

Change-Id: I5e52f2b80b93d564b12f42cc2ed6c0f10e9b03cc
Signed-off-by: Eric Miao <emiao@nvidia.com>