Update patch set 2
Patch Set 2:

(2 comments)

In unit testing, one command needs to run multiple times, it has to initialize and clean up each time. The code is simplified if the command instance is created each time it is needed. And we don't need xxxCommandImpl classes any more. Some comments about the change are added in the commit line.

Patch-set: 2
Label: Verified=0