Update patch set 1
Patch Set 1:

It is the event counts hit on this item / total event counts. It is not the same as SamplePercent. Between two samples, there can have more than one event counts, which depends on the sample freq/period.I don't think "Overhead" is a very good name, but perf uses it. So I guess we can think of it as the time spent in one item when using cpu-cycle event.

Patch-set: 1
Label: Verified=0