Update patch set 1
Patch Set 1:

> AOSP uses clang for host/target by default, and only uses g++ for
 > test on target. So I think there is no need to keep USE_CLANG flag
 > in Android.mk.
 > Previous code initializes struct members with designation, which is
 > not part of c++ standard.
 > It shows as error in eclipse IDE.

 > > AOSP uses clang for host/target by default, and only uses g++ for
 > > test on target. So I think there is no need to keep USE_CLANG
 > flag
 > > in Android.mk.
 > 
 > AOSP uses clang for host and shamu, but most target modules still
 > default to clang.

what danalbert meant to say was "AOSP uses clang for host and shamu, but most target modules still default to GCC". that is: your change means that we'd start building the non-shamu target simpleperf binary with GCC, which we don't want to do.

we're trying to move away from GCC completely. one of the reasons is that we always have to support clang anyway, and life is easier if you only have to support one compiler. this patch just makes our lives worse.

Patch-set: 1