Update patch set 3
Patch Set 3:

If we had an alternative implementation of the Reed-Solomon primitives that are compatible with the kernel, we wouldn't need external/fec. It looks like that would mean writing them from scratch though.

This library uses the primitives in external/fec to implement proper error correction with interleaving and support for locating erasures using verity hashes.

Patch-set: 3
Label: Verified=0