Tool to replay memory allocation recordings.
Included with the tool are a set of app memory recordings that
can be used to judge the abilities of the native memory
allocator.

Verified the full flow by running valgrind on the host version on all
the current dumps and verifying that there are no leak, and no errors.

Change-Id: I756224991f04b61ba26c692a30a37aad0c9fb163