Update patch set 5
Patch Set 5:

This file also needs to be called *_test.py to make it actually run the test.

I just tried this on the Dragon 3.18 kernel and it doesn't seem to crash the kernel after running it a handful of times... Was it expected to always crash the kernel?

We do have a few extra patches to tcp_nuke_addr in our tree, namely:
https://chromium-review.googlesource.com/284953
https://chromium-review.googlesource.com/284954

I'll test it on the android-3.10 tree as well.

Cheers,
Filipe

Patch-set: 5