Update patch set 7
Patch Set 7:

> IPV4_LOOPBACK_ADDR is currently unused. I will send a patch with a
 > corresponding IPv4 test after this one gets merged.

Hi Subash,

In that case, I reworked it to use generic names on the functions and to take an additional addr_family argument.

To add IPv4 support, all you should need is to define the correct ifreq for it in KillAddrIoctl and then add a testIPv4KillAddr that does the same but passing AF_INET instead of AF_INET6.

In the process, I also used contextlib.closing() to call close() on all the sockets and defined a RunInBackground() to run the thread on a context manager and make sure we wait for it to terminate when the block is over.

Please take another look.

Cheers!
Filipe

Patch-set: 7