Update patch set 1
Patch Set 1:

We were having issues with developers devices not booting. The problem could be reproduced by something like:

0. Start with a device with verity enabled on /system
1. adb disable-verity
2. remount system rw
3. make some change to the fs
4. hard crash (without unmounting /system)

When the device tried to reboot fs_mgr use ext4_parse_sb to figure out the size of the block device in order to find verity metadata. The ext4_parse_sb would however fail since the super block of says the file system was not unmounted cleanly. This left the device lost.
It is fine to parse out the size of the block device even if the file system was not unmounted cleanly, so it should
be ok to simply remove the test.

The only other client of ext4_parse_sb is ext4_parse_sb_info and it is explicitly checking s_state so it should be unaffected by this change.

Patch-set: 1