Update patch set 1
Patch Set 1: Code-Review-1

Holding off committing until our group is satisfied that this should be the home for kernel functionality gTests.

Carries some technical debt because some of the tests in these directories may need to be converted to gTests before we can notify cts to grab them for their regression testing.

b/26559308 opened to deal with this technical debt or a decision as to where to house kernel gTests.

Patch-set: 1