Update patch set 1
Patch Set 1: Code-Review+2

Discussions appear to accept that the current and future kernel tests will remain somewhat scattered. There will be tests with their own repositories in the future. Documenting where they are is the engineering effort and this is already a place for several of them.

Patch-set: 1
Label: Code-Review=+2