Update patch set 1
Patch Set 1:

> I believe that Bahdri had some plans to change this earlier.
 > 
 > Out of curiosity, on which host/target architecture combination do
 > you see the problem? Compiling with clang on x86_64 and running on
 > 32-bit ARMv7 appears to work fine at least.

I caught the issue on a x86 Android device.  Also, on my host (x86), using gcc compiler, if I compile for 64 bits the structure is 64 bytes large while if I compile for 32 bits the structure is 60 bytes large. Based on the structure definition it all makes sense to me.

Patch-set: 1