Update patch set 3
Patch Set 3: Code-Review-1

I've run into a problem while testing this patch. On some perf.data files the quipper parsing code will fail to map all of the IPs for a given callchain: when this happens the length of the callchain reported by the parser is zero, which triggers an assert in the converter (makes it look as though there is a mix of flat profile + callchain profile). I need to revise the converter code to account for this; will post another patch later today.

Patch-set: 3
Label: Code-Review=-1