Update patch set 1
Patch Set 1:

(3 comments)

From the perspective of a client reading the proto, it seems as though it would be cleaner to just read a single message that has a series of Records, e.g.

message Toplevel {
  ...
  repeated Record = 1;
}

However I am guessing that you want avoid the need to buffer all of the sample records in simpleperf and do a single write at the end of the run (this makes sense-- there is a strong incentive to emit records on the fly as opposed to buffering everything).

If this is the case, I think the code would be cleaner if you just used something like "SerializeWithCachedSizesToArray" to encode each record to an array of bytes and then write the result out into a file (using your existing scheme of recording the size for each record) as opposed to calling into the CodedOutputStream/CodedInputStream classes directly. Things like "CodedInputStream::PushLimit" seem veruy low level -- I would worry that your code would break as the protobuf implementation evolves/changes.

If you're comfortable using the lower-level classes / interfaces, though, then it looks good overall to me.

Patch-set: 1
Reviewer: Gerrit User 1059614 <1059614@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0