Update patch set 2
Patch Set 2:

(3 comments)

Yes, the reason to not use Toplevel message is I don't want to buffer everything before writing. And protobuf seems not indend to build large messages. The main reason to use CodedInputStream and CodedOutputStream is to read/write LittleEndian32 message size. As PushLimit is public member of CodedOutputStream, I think it is fine to use.

Patch-set: 2
Label: Verified=0