Update patch set 1
Patch Set 1:

I wonder if symbolizing on the fly is the most efficient approach. It involves a significant performance penalty. Have you considered adding a mini symbol tables at the end of perf.data which can be generated at the end of the profiling process. This would not impact the sampling but still allow report to generate a symbolized report without access to dso files.

Patch-set: 1