Update patch set 1
Patch Set 1:

CTS tests runs in app domain if the test is an app. But simpleperf cts test is a native test, and I don't see any hint that cts wraps native test in an app, so probably it still runs in shell domain. I will build cts test and verify this.

Patch-set: 1
Label: Verified=0