Update patch set 10
Patch Set 10: Code-Review-1

This patch, as near as I can tell, makes to changes to how the ext4 file system is created using make_ext4fs.  #1: It makes changes to track how blocks are allocated and freed in make_ext4fs (which shouldn't really make a difference).   #2:  It sets the reserved blocks field in the superblock, which had formerly to be zero, a calculated value.    The change in #2, however, has nothing to do with the change in the referenced kernel commit.  There is a collision in "reserved blocks" named concept.   The reserved blocks in the superblock refers to blocks which are reserved for use by the root user (or the root gid, depending on how the superblock is configured).  This is described in the tune2fs(8) and mke2fs(8) man pages.   The reserved blocks concept as added by Jan Kara is a different thing.  It is not something which is adjusted via a superblock setting, but rather reserved 2% of the file system, or 4096 blocks (whichever is smaller) for use only by metadata blocks (e.g., indirect blocks or extent tree blocks).

So I very much doubt this patch actually accomplished anything.  I suspect what happened is this got worked around in AOSP around about the same time, which was to simply add extra free space to the created system image to prevent this scenario from happening.   (Which if I recall correctly, happened when trying to replace an existing file when using a file-based system update mechanism.)   I suspect if you revert this change, you'll find that it doesn't change any observable behavior.

Patch-set: 10
Reviewer: Gerrit User 1038033 <1038033@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1