Reduce memory usage of libufdt
This commit slightly modifies the data struture for unflattened device
tree introduced in previous commit. The main difference is replacing
the hash table with a linked list for maintaining children of a device
tree node.

The performance is significantly improved in practice. For device tree
overlay including two trees (2500 and 500 nodes, respectively), both
running time and additional memory usage is halved: 19 ms -> 10 ms and
350 KB -> 282 KB, respectively.

Bug: 30800619

Change-Id: I2a08f179cce3040a7b093656324e29138ff5fc66