Update patch set 1
Patch Set 1:

This appears to be coming from an external repo -- it at least needs information about where it came from. But shouldn't this be in new repo in external/... in that case? It would be good to clarify which parts of this change came from the upstream source, and which were local modifications (easiest by splitting into two changes).

You're missing the NOTICE file required by the headers (which are partially missing -- are they missing upstream too?)

Patch-set: 1