Update patch set 1
Patch Set 1:

We generally prefer that any "external" libraries pulled in at least will be able to be updated in the future. Importing such a library requires owning the future maintenance of that component too, as there are often security backports, etc. that might be needed. There really should be a maintenance plan for any external code being imported here.

Patch-set: 1