Update patch set 1
Patch Set 1: Code-Review+1

If I want to check it in? Sure :-).

If you want I can adapt it, there are some topics which I was not sure.
One is that I did keep the existing C API, hence added a new set of all the functions with the added instance pointers.
That is the currect way if we need to maintain the compatibility. I'm not sure if that is necessary, if not I can clean it up, add the lib reference to the existing functions
and that would avoid some of the API duplication. Up to you if I should do that.

The benefit of the current implementation is that it does allow to process multiple perf.data files in one session. Previously stale data in the single instance did break that,
I found that quite problematic in some tests I wrote. There is still the issue that the library code itself uses some globals, but that was not an issue for my usage.
Basically I had the same kernel/etc in all the usages anyhow. Also if we want to address that in the future than that needs this change regardless.

So I'm all for including it. Let me know if you want me to change the API duplication mentioned above.
Thanks for looking into it.

Dainel

Patch-set: 1
Label: Code-Review=+1