Update patch set 2
Patch Set 2:

(1 comment)

Cleaned up the copies, now the python API remains stable, but the C API of libsimpleperf_report.so is not.
I think this is more reasonable.
I did replace the checked in libsimpleperf_report.so as the API from the previous one did change.
I personally would not add such large files into revision control, but if one is contained, it does need to be updated.

Patch-set: 2