Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> i think this is right; you'll get libc++ by default because you didn't say you _didn't_ want an STL. dwillemsen?

Correct, this is automatically getting a copy of libc++_static since LOCAL_CXX_STL isn't set to 'none'. So this is definitely a good change, but why is it enough?

I'd guess this wasn't a problem before because none of the conflicting symbols were being required (even transitively) from the binary. But if that changed, why doesn't it also need a reference to libbase now?

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>