add ziparchive-perf benchmark
This patch adds two benchmarks that measure the performance of some operations
of libziparchive.

Both benchmarks are creating a temporary zip file containing file names of
uniformly distributed lengths.  The creation of the zip file is not timed
in the benchmarks.

- findIt tries to find an inexisting entry in the files of the zip archive,
in order to force the code to examine all the files in the archive.

- iterateAll uses the iterate function to list all the files in the archive.

To build the benchmark,

$ cd aosp
$ source build/envsetup.sh
$ lunch aosp_hikey-eng
$ cd system/extras/ziparchive-perf
$ mma tests

The benchmark gets installed in the following dir:
out/target/product/hikey/data/benchmarktest64/ziparchivePerf/ziparchivePerf64
The benchmark is installed in the data.img and can be flashed on the device.
One can also copy the benchmark and execute it on aarch64-linux, as it is
statically linked.

To execute only one of the benchmarks:

$ ./ziparchivePerf64 --benchmark_filter=findIt

Also see --help for more options on how to run the benchmarks.  It is also
useful on a big/little machine to use taskset to run the benchmark on a specific
processor.  To select an A57 processor on a Juno aarch64-linux machine:

$ taskset -c 1 ./ziparchivePerf64 --benchmark_filter=findIt --benchmark_repetitions=3
[...]
Benchmark                  Time           CPU Iterations
--------------------------------------------------------
findIt               2389668 ns    2389632 ns        293
findIt               2389449 ns    2388866 ns        293
findIt               2389651 ns    2389097 ns        293
findIt_mean          2389589 ns    2389198 ns        293
findIt_stddev             99 ns        321 ns          0

Change-Id: I52f94238a8dab5f4dbcf868a90edde0e7c9fef61