Update patch set 2
Patch Set 2:

So there is no chance that a device with a kernel that doesn't recognize INCOMPAT_ENCRYPT will be updated to the latest version of Android?

Also, the 'fastboot -w' and 'fastboot format' commands create a filesystem image on the host, then flash it to the device.  Assuming that a given 'fastboot' version might be used on different devices of different ages, how do we make it set the correct flag?

Patch-set: 2
Reviewer: Gerrit User 1137063 <1137063@85c56323-6fa9-3386-8a01-6480fb634889>