simpleperf: change the way running cts tests.
Before this CL, CtsSimpleperfTestCases copies itself to the app's
directory, then run it using run-as.

With this CL, CtsSimpleperfTestCases keeps itself in /data/local/tmp,
but forces stat/record cmd to run with --app option. This gives
more freedom to tests:
1. They can stay in shell's context with --in-app option.
2. The stat/record cmds are started in the shell's context, so
they can collect information no available in app's context (like
data in /sys/kernel/debug/tracing/events).

This is a preparation to add tests for recording tracepoint events.
It also matches the way we want users to use simpleperf (with --app option).

Bug: http://b/29520177
Test: run CtsSimpleperfTestCases.

Change-Id: I1709adfb1ff7169df87560226c197e473fdf8516