Update patch set 1
Patch Set 1:

It works as (1), this field is populated in each sample if trace_offcpu = true.
Yes, it is redundant and can be calculated by time field. If the android studio profiler would like to calculate this, then we don't need to add time_to_next_sample field. But make sure that it is time_to_next_sample, not time_from_previous_sample. Because we get a sample when a thread is scheduled out, not scheduled in.

Do we need to know which callchain is for sleeping?
With trace_offcpu = true, we dump a sample in two cases: one is a thread has run enough cpu-cycle events to hit a sample, second is when a thread is scheduled out. This is already enough for us to know where we are spending most of the time, whether it is sleeping or running.

But since we only know when a thread is scheduled out, not when it is scheduled in, we don't know strictly how long a thread is sleeping on a callchain. I guess the main reason you want to distinguish sleep callchain is to to show strictly sleeping time and callchain. But this is another feature that hasn't been implemented. So no need to add interface to this. Let me know if you are also interested in this, I can make it a higher priority.

Patch-set: 1
Reviewer: Gerrit User 1056364 <1056364@85c56323-6fa9-3386-8a01-6480fb634889>