Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> you'll need to add a NOTICE symlink to LICENSE.txt to ensure that the license gets collected up properly for the NDK, i think. +danalbert.

It won't be picked up automatically either way (there's an error if you have nothing, but if there's a NOTICE in the root of the package, then it's assumed that's the correct license). You'll need to either merge them here or adjust the build definition in ndk/checkbuild.py to make sure it gets included.

Patch-set: 2
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>