Update patch set 1
Patch Set 1: Code-Review-1

it's not really houdini though... we'd want this to apply to ndk_translation or any other equivalent "i'm running an abi that doesn't match the cpu".

can we tell from uname(2) or /proc/cpuinfo when we're being lied to? presumably not, because presumably these things try to fake it everywhere?

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1