simpleperf: windows dependency -lws2_32
Add

    LOCAL_LDLIBS_windows := -lws2_32

to Android.mk.  We add a few extra to the linux-only host builds to
head off any future expansion to windows deliverables.  This keeps on
having to come back during admittedly abandoned prototyping for new
features so let's make it permanent to allow delivery or
experimentation in the dependent libraries.

Test: compile
Bug: 63736262
Change-Id: I6de4783c6c48f420b63d6c1392e3de190094ec04