Update patch set 1
Patch Set 1:

If the main thread is the place where most of the business happens and takes the most cpu time, then after this patch, and the main thread is still at the top, and other threads are sorted by how much business happen in those threads. Why don't you try it to check if it helps?

Patch-set: 1
Reviewer: Gerrit User 1056364 <1056364@85c56323-6fa9-3386-8a01-6480fb634889>